+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/constants/UrlParameterConstants.kt
@@ -1,0 +1,3 @@
+package uk.gov.communities.prsdb.webapp.constants
+
+const val CONTEXT_ID_URL_PARAMETER = "contextId"
