+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/controllers/LandlordController.kt
@@ -1,11 +1,20 @@
 package uk.gov.communities.prsdb.webapp.controllers
 
+import jakarta.validation.Valid
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
+import org.springframework.validation.BindingResult
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.ModelAttribute
+import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.util.UriTemplate
+import uk.gov.communities.prsdb.webapp.constants.BACK_URL_ATTR_NAME
+import uk.gov.communities.prsdb.webapp.constants.CONTEXT_ID_URL_PARAMETER
 import uk.gov.communities.prsdb.webapp.constants.DASHBOARD_PATH_SEGMENT
+import uk.gov.communities.prsdb.webapp.constants.DELETE_INCOMPLETE_PROPERTY_PATH_SEGMENT
 import uk.gov.communities.prsdb.webapp.constants.INCOMPLETE_PROPERTIES_PATH_SEGMENT
 import uk.gov.communities.prsdb.webapp.constants.LANDLORD_DETAILS_PATH_SEGMENT
 import uk.gov.communities.prsdb.webapp.constants.LANDLORD_PATH_SEGMENT
@@ -15,6 +24,8 @@ import uk.gov.communities.prsdb.webapp.constants.RENTERS_RIGHTS_BILL_URL
 import uk.gov.communities.prsdb.webapp.controllers.LandlordController.Companion.LANDLORD_BASE_URL
 import uk.gov.communities.prsdb.webapp.exceptions.PrsdbWebException
 import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
+import uk.gov.communities.prsdb.webapp.models.requestModels.formModels.DeleteIncompletePropertyRegistrationAreYouSureFormModel
+import uk.gov.communities.prsdb.webapp.models.viewModels.formModels.RadiosButtonViewModel
 import uk.gov.communities.prsdb.webapp.models.viewModels.summaryModels.IncompletePropertiesViewModel
 import uk.gov.communities.prsdb.webapp.services.LandlordService
 import uk.gov.communities.prsdb.webapp.services.PropertyRegistrationService
@@ -83,9 +94,84 @@ class LandlordController(
         return "incompletePropertiesView"
     }
 
+    @GetMapping("/$DELETE_INCOMPLETE_PROPERTY_PATH_SEGMENT")
+    fun deleteIncompletePropertyAreYouSure(
+        model: Model,
+        principal: Principal,
+        @RequestParam(value = "contextId", required = true) contextId: Long,
+    ): String {
+        populateDeleteIncompletePropertyRegistrationModel(model, contextId, principal.name)
+        model.addAttribute(
+            "deleteIncompletePropertyRegistrationAreYouSureFormModel",
+            DeleteIncompletePropertyRegistrationAreYouSureFormModel(),
+        )
+
+        return "deleteIncompletePropertyRegistration"
+    }
+
+    @PostMapping("/$DELETE_INCOMPLETE_PROPERTY_PATH_SEGMENT")
+    fun deleteIncompletePropertyAreYouSure(
+        model: Model,
+        principal: Principal,
+        @RequestParam(value = "contextId", required = true) contextId: Long,
+        @Valid
+        @ModelAttribute
+        formModel: DeleteIncompletePropertyRegistrationAreYouSureFormModel,
+        bindingResult: BindingResult,
+    ): String {
+        if (bindingResult.hasErrors()) {
+            populateDeleteIncompletePropertyRegistrationModel(model, contextId, principal.name)
+            return "deleteIncompletePropertyRegistration"
+        }
+
+        if (formModel.wantsToProceed == true) {
+            propertyRegistrationService.deleteIncompleteProperty(contextId.toLong(), principal.name)
+        }
+
+        return "redirect:$INCOMPLETE_PROPERTIES_URL"
+    }
+
+    fun populateDeleteIncompletePropertyRegistrationModel(
+        model: Model,
+        contextId: Long,
+        principalName: String,
+    ) {
+        val formContext =
+            propertyRegistrationService.getIncompletePropertyFormContextForLandlordIfNotExpired(
+                contextId,
+                principalName,
+            )
+        val singleLineAddress = propertyRegistrationService.getAddressData(formContext).singleLineAddress
+
+        model.addAttribute(
+            "radioOptions",
+            listOf(
+                RadiosButtonViewModel(
+                    value = true,
+                    valueStr = "yes",
+                    labelMsgKey = "forms.radios.option.yes.label",
+                ),
+                RadiosButtonViewModel(
+                    value = false,
+                    valueStr = "no",
+                    labelMsgKey = "forms.radios.option.no.label",
+                ),
+            ),
+        )
+        model.addAttribute("singleLineAddress", singleLineAddress)
+        model.addAttribute(BACK_URL_ATTR_NAME, INCOMPLETE_PROPERTIES_URL)
+    }
+
     companion object {
         const val LANDLORD_DASHBOARD_URL = "/$LANDLORD_PATH_SEGMENT/$DASHBOARD_PATH_SEGMENT"
         const val LANDLORD_BASE_URL = "/$LANDLORD_PATH_SEGMENT"
         const val INCOMPLETE_PROPERTIES_URL = "/$LANDLORD_PATH_SEGMENT/$INCOMPLETE_PROPERTIES_PATH_SEGMENT"
+
+        const val DELETE_INCOMPLETE_PROPERTY_ROUTE =
+            "/$LANDLORD_PATH_SEGMENT/$DELETE_INCOMPLETE_PROPERTY_PATH_SEGMENT" +
+                "?$CONTEXT_ID_URL_PARAMETER={contextId}"
+
+        fun deleteIncompletePropertyPath(contextId: Long): String =
+            UriTemplate(DELETE_INCOMPLETE_PROPERTY_ROUTE).expand(contextId).toASCIIString()
     }
 }
