+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/DeleteIncompletePropertyRegistrationAreYouSurePage.kt
@@ -1,9 +1,9 @@
 package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages
 
 import com.microsoft.playwright.Page
+import uk.gov.communities.prsdb.webapp.constants.CONTEXT_ID_URL_PARAMETER
 import uk.gov.communities.prsdb.webapp.constants.DELETE_INCOMPLETE_PROPERTY_PATH_SEGMENT
-import uk.gov.communities.prsdb.webapp.constants.REGISTER_PROPERTY_JOURNEY_URL
-import uk.gov.communities.prsdb.webapp.controllers.RegisterPropertyController.Companion.CONTEXT_ID_URL_PARAMETER
+import uk.gov.communities.prsdb.webapp.constants.LANDLORD_PATH_SEGMENT
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Heading
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.AreYouSureFormBasePage
 
@@ -12,7 +12,7 @@ class DeleteIncompletePropertyRegistrationAreYouSurePage(
     urlArguments: Map<String, String>,
 ) : AreYouSureFormBasePage(
         page,
-        "/$REGISTER_PROPERTY_JOURNEY_URL/$DELETE_INCOMPLETE_PROPERTY_PATH_SEGMENT" +
+        "/$LANDLORD_PATH_SEGMENT/$DELETE_INCOMPLETE_PROPERTY_PATH_SEGMENT" +
             "?$CONTEXT_ID_URL_PARAMETER=${urlArguments["contextId"]}",
     ) {
     val heading = Heading(page.locator("h1.govuk-heading-l"))
